This is a much larger commit than intended. Here's a rough breakdown:
* Re-architected expectation builders to more closely match the way
jMock works: LRExpectations is the main builder, and the old LRExpectationBuilder
is now LRInvocationExpectationBuilder and just builds a single expectation.

* Started to simplify LRInvocationExpectation, added unit tests, and rather
than have it use an NSInvocation when matching, supply it with target/selector
and parameters separately.

* LRInvocationExpectation uses a matcher for parameters. 

* OCHamcrest is now a project dependency. It just made sense.

* LRExpectations handles notification expectations, not the mockery. Its
just a different type of expectation. Code still needs removing from the
mockery.

* Invocations on mocks outside of the upper bounds of its cardinality 
will just trigger unexpected invocation errors. 

* Removed neverReceives, as all mocks will raise errors if they receive
a message they shouldn't - the behaviour is implicit.